+++ test/components/network_analyses/test_detour_artifact.py
@@ -45,11 +45,21 @@ def test_apply_detour_color_and_label(default_polygon_geometry):
 
     expected_detour_df = gpd.GeoDataFrame(
         {
-            'detour_factor': [2.5, 3.5, np.nan],
-            'geometry': [default_polygon_geometry, default_polygon_geometry, default_polygon_geometry],
-            'detour_category': [DetourCategory.MEDIUM_DETOUR, DetourCategory.HIGH_DETOUR, DetourCategory.UNREACHABLE],
-            'color': [Color('#eea321'), Color('#e75a13'), Color('#990404')],
-            'label': ['Medium Detour', 'High Detour', 'Unreachable'],
+            'detour_factor': [1.0, 2.5, 3.5, np.nan],
+            'geometry': [
+                default_polygon_geometry,
+                default_polygon_geometry,
+                default_polygon_geometry,
+                default_polygon_geometry,
+            ],
+            'detour_category': [
+                DetourCategory.LOW_DETOUR,
+                DetourCategory.MEDIUM_DETOUR,
+                DetourCategory.HIGH_DETOUR,
+                DetourCategory.UNREACHABLE,
+            ],
+            'color': [Color('#FFFFE0'), Color('#eea321'), Color('#e75a13'), Color('#808080')],
+            'label': ['Low Detour', 'Medium Detour', 'High Detour', 'Unreachable'],
         },
         crs=CAN_DEFAULT_CRS,
     )
@@ -62,25 +72,37 @@ def test_summarise_detour(default_polygon_geometry):
     input_hexgrid = gpd.GeoDataFrame(
         data={
             'detour_factor': [0, 3, 6, 10],
+            'color': [Color('#FFFFE0'), Color('#eea321'), Color('#e75a13'), Color('#e75a13')],
+            'label': ['Low Detour', 'Medium Detour', 'High Detour', 'High Detour'],
             'geometry': 4 * [default_polygon_geometry],
         },
         crs=CAN_DEFAULT_CRS,
     )
-    chart = summarise_detour(hexgrid=input_hexgrid)
+
+    chart = summarise_detour(detour_factor_data=input_hexgrid)
 
     assert isinstance(chart, go.Figure)
-    np.testing.assert_array_equal(chart['data'][0]['x'], ([0, 3, 6, 10]))
+    np.testing.assert_array_equal(
+        chart['data'][0]['x'], (['Low Detour (0 to 1.99)', 'Medium Detour (2.0 to 2.99)', 'High Detour (>= 3)'])
+    )
 
 
 def test_summarise_detour_inf(default_polygon_geometry):
     input_hexgrid = gpd.GeoDataFrame(
         data={
-            'detour_factor': [0, 3, 6, np.inf],
-            'geometry': 4 * [default_polygon_geometry],
+            'detour_factor': [2.5, 3.5, np.nan, 0, np.nan],
+            'color': [Color('#eea321'), Color('#e75a13'), Color('#808080'), Color('#FFFFE0'), Color('#808080')],
+            'label': ['Medium Detour', 'High Detour', 'Unreachable', 'Low Detour', 'Unreachable'],
+            'geometry': 5 * [default_polygon_geometry],
         },
         crs=CAN_DEFAULT_CRS,
     )
-    chart = summarise_detour(hexgrid=input_hexgrid)
+
+    chart = summarise_detour(detour_factor_data=input_hexgrid)
 
     assert isinstance(chart, go.Figure)
-    np.testing.assert_array_equal(chart['data'][0]['x'], ([0, 3, 6, np.inf]))
+    np.testing.assert_array_equal(
+        chart['data'][0]['x'],
+        (['Low Detour (0 to 1.99)', 'Medium Detour (2.0 to 2.99)', 'High Detour (>= 3)', 'Unreachable']),
+    )
+    np.testing.assert_array_equal(chart['data'][0]['y'], ([20, 20, 20, 40]))