+++ test/components/path_lighting/test_path_lighting_analysis.py
@@ -1,5 +1,6 @@
 import geopandas as gpd
 import geopandas.testing
+import pandas as pd
 
 from walkability.components.path_lighting.path_lighting_analysis import (
     get_path_lighting,
@@ -33,7 +34,7 @@ def test_get_path_lighting(default_path_geometry, default_polygon_geometry):
             'osm_tags': [{'lit': 'yes'}, {'lit': 'automatic'}, {}],
             'geometry': [default_path_geometry, default_path_geometry, default_polygon_geometry],
             'path_lighting': [PathLightingCategory.YES, PathLightingCategory.AUTOMATIC, PathLightingCategory.UNKNOWN],
-            'path_lighting_rating': [1, 0.8, None],
+            'path_lighting_rating': pd.Series([1, 0.8, None]),
         },
         crs=CAN_DEFAULT_CRS,
     )