+++ walkability/components/network_analyses/detour_analysis.py
@@ -3,9 +3,8 @@ from enum import Enum
 from pathlib import Path
 
 import geopandas as gpd
-import matplotlib.colors as mcolors
-import matplotlib.pyplot as pyplt
 import numpy as np
+import pandas as pd
 import plotly.graph_objects as go
 import shapely
 from climatoology.base.artifact_creators import (
@@ -40,7 +39,8 @@ def detour_factor_analysis(
     hexcell_artifact = build_detour_factor_artifact(detour_factor_data=detour_factors, resources=resources)
 
     summary = summarise_detour(detour_factors)
-    summary_artifact = build_detour_summary_artifact(summary, resources=resources)
+    n_inf = sum(np.isinf(detour_factors['detour_factor']))
+    summary_artifact = build_detour_summary_artifact(summary, n_inf, resources=resources)
 
     return [hexcell_artifact, summary_artifact]
 
@@ -51,6 +51,7 @@ def build_detour_factor_artifact(
     """Artifact containing a GeoJSON with hex-grid cells and the Detour Factor."""
 
     data = apply_color_and_label(detour_factor_data, cmap_name)
+    data = data[data['detour_category'] != DetourCategory.LOW_DETOUR]
 
     return create_vector_artifact(
         data=data[['index', 'geometry', 'detour_factor', 'color', 'label']],
@@ -74,23 +75,22 @@ class DetourCategory(Enum):
 
 def apply_color_and_label(detour_factor_data: gpd.GeoDataFrame, cmap_name: str = 'YlOrRd') -> gpd.GeoDataFrame:
     def categorize_detour(detour_value):
-        if detour_value < DetourCategory.MEDIUM_DETOUR.value:
+        if np.isinf(detour_value) or pd.isna(detour_value):
+            return DetourCategory.UNREACHABLE
+        elif detour_value < DetourCategory.MEDIUM_DETOUR.value:
             return DetourCategory.LOW_DETOUR
         elif detour_value < DetourCategory.HIGH_DETOUR.value:
             return DetourCategory.MEDIUM_DETOUR
-        elif detour_value >= DetourCategory.HIGH_DETOUR.value:
-            return DetourCategory.HIGH_DETOUR
         else:
-            return DetourCategory.UNREACHABLE
+            return DetourCategory.HIGH_DETOUR
 
     detour_factor_data['detour_category'] = detour_factor_data['detour_factor'].apply(categorize_detour)
-    detour_factor_data_no_low = detour_factor_data[detour_factor_data['detour_category'] != DetourCategory.LOW_DETOUR]
 
-    detour_factor_data_no_low['color'] = detour_factor_data_no_low['detour_category'].map(DETOUR_FACTOR_COLOR_MAP)
+    detour_factor_data['color'] = detour_factor_data['detour_category'].map(DETOUR_FACTOR_COLOR_MAP)
 
-    detour_factor_data_no_low['label'] = detour_factor_data_no_low.detour_category.apply(apply_labels)
+    detour_factor_data['label'] = detour_factor_data.detour_category.apply(apply_labels)
 
-    return detour_factor_data_no_low
+    return detour_factor_data
 
 
 def apply_labels(detour_category: DetourCategory) -> str:
@@ -99,12 +99,13 @@ def apply_labels(detour_category: DetourCategory) -> str:
             return 'Medium Detour'
         case DetourCategory.HIGH_DETOUR:
             return 'High Detour'
+        case DetourCategory.LOW_DETOUR:
+            return 'Low Detour'
         case DetourCategory.UNREACHABLE:
             return 'Unreachable'
 
 
-def build_detour_summary_artifact(aoi_aggregate: go.Figure, resources: ComputationResources) -> Artifact:
-    n_inf = sum(np.isinf(aoi_aggregate['data'][0]['x']))
+def build_detour_summary_artifact(aoi_aggregate: go.Figure, n_inf: int, resources: ComputationResources) -> Artifact:
     return create_plotly_chart_artifact(
         figure=aoi_aggregate,
         metadata=ArtifactMetadata(
@@ -120,41 +121,50 @@ def build_detour_summary_artifact(aoi_aggregate: go.Figure, resources: Computati
 
 
 def summarise_detour(
-    hexgrid: gpd.GeoDataFrame,
+    detour_factor_data: gpd.GeoDataFrame,
 ) -> go.Figure:
     log.info('Summarising detour factor stats')
-    stats = hexgrid.dropna(how='any')
 
-    min_value = stats['detour_factor'].min()
-    max_value = stats.loc[stats['detour_factor'] != np.inf, 'detour_factor'].max()
-    counts, bin_edges = np.histogram(stats['detour_factor'], bins=30, range=(min_value, max_value))
-
-    cmap = pyplt.get_cmap('YlOrRd', len(counts))
-    colors = [mcolors.to_hex(cmap(i)) for i in range(len(counts))]
-
-    histogram = go.Histogram(
-        x=stats['detour_factor'],
-        nbinsx=30,
-        histnorm='percent',
-        marker=dict(color=colors),
-        hovertemplate='Range: %{x}<br>Percentage: %{y:.2f}%<extra></extra>',
+    # Add detour factor categories for labels
+    detour_factor_range = {
+        'Low Detour': 'Low Detour (0 to 1.99)',
+        'Medium Detour': 'Medium Detour (2.0 to 2.99)',
+        'High Detour': 'High Detour (>= 3)',
+        'Unreachable': 'Unreachable',
+    }
+    detour_factor_data['label'] = detour_factor_data['label'].map(detour_factor_range)
+    detour_factor_data = detour_factor_data.sort_values(
+        by='detour_factor',
+        ascending=True,
     )
 
-    detour_fig = go.Figure(data=[histogram])
+    stats = detour_factor_data.groupby(['label', 'color'], sort=False).size().reset_index(name='count')
+    totals = stats['count'].sum()
+    colors = stats['color']
 
-    detour_fig.update_layout(
+    bar_fig = go.Figure(
+        data=go.Bar(
+            x=stats['label'],
+            y=(stats['count'] / totals) * 100,
+            marker_color=[c.as_hex() for c in colors],
+            hovertemplate='Range: %{x}<br>Percentage: %{y:.2f}%<extra></extra>',
+        )
+    )
+
+    bar_fig.update_layout(
         title=dict(
             subtitle=dict(text='Percentage', font=dict(size=14)),
         ),
-        xaxis_title='Detour Factor',
+        xaxis_title='Detour Class',
         yaxis_title=None,
         margin=dict(t=30, b=60, l=80, r=30),
     )
-    return detour_fig
+    return bar_fig
 
 
 DETOUR_FACTOR_COLOR_MAP = {
+    DetourCategory.LOW_DETOUR: Color('#FFFFE0'),
     DetourCategory.MEDIUM_DETOUR: Color('#eea321'),
     DetourCategory.HIGH_DETOUR: Color('#e75a13'),
-    DetourCategory.UNREACHABLE: Color('#990404'),
+    DetourCategory.UNREACHABLE: Color('#808080'),
 }